+++ source/client/scene/controller/RenderingController.cpp
@@ -52,7 +52,7 @@ void RenderingController::update(entt::registry &registry) {
 void RenderingController::draw(entt::registry &registry, gk::RenderTarget &target, gk::RenderStates states) {
 	registry.view<DrawableComponent, PositionComponent, RotationComponent>().each([&](auto, auto &drawable, auto &position, auto &rotation) {
 		gk::Transformable transformable;
-		transformable.setPosition(position.x, position.y, position.z);
+		transformable.setPosition((float)position.x, (float)position.y, (float)position.z);
 		transformable.getRotationTransform().getMatrix() = glm::toMat4(rotation.quat);
 
 		gk::RenderStates drawStates = states;
